Implement CanBuildWithDefault
