+++ crates/cgp-field/src/traits/transform_type.rs
@@ -1,4 +1,4 @@
-use crate::{IsNothing, IsPresent, MapType};
+use crate::MapType;
 
 pub trait TransformMapFields<Transform, TargetMap> {
     type Output;
@@ -9,18 +9,4 @@ pub trait TransformMapFields<Transform, TargetMap> {
 /// Natural transformation from M1::Map<T> to M2::Map<T>
 pub trait TransformMap<M1: MapType, M2: MapType, T> {
     fn transform_mapped(value: M1::Map<T>) -> M2::Map<T>;
-}
-
-pub struct TransformMapDefault;
-
-impl<T> TransformMap<IsPresent, IsPresent, T> for TransformMapDefault {
-    fn transform_mapped(value: T) -> T {
-        value
-    }
-}
-
-impl<T: Default> TransformMap<IsNothing, IsPresent, T> for TransformMapDefault {
-    fn transform_mapped(_value: ()) -> T {
-        T::default()
-    }
 }
